refactor(keychain): relocate app secrets store from pkg/auth
Avoids a config<->auth import cycle: the upcoming read-resolution step (pkg/config) needs LoadAppSecrets, but pkg/auth already imports pkg/config. A neutral pkg/keychain package both can depend on.
